+++ sphinxcontrib/osint/plugins/__init__.py
@@ -95,6 +95,10 @@ class PluginSource(Plugin):
         return None
 
     @classmethod
+    def local(cls, directive, source_name):
+        return None
+
+    @classmethod
     def init_source(cls, env, osint_source):
         pass