+++ sphinxcontrib/osint/scripts/text.py
@@ -10,30 +10,26 @@ import os
 import sys
 from datetime import date
 import json
+import click
 
 from sphinx.application import Sphinx
 from sphinx.util.docutils import docutils_namespace
 
 from ..plugins.text import Text
-from . import parser_makefile, get_parser
+from . import parser_makefile, cli
 
 
 __author__ = 'bibi21000 aka Sébastien GALLET'
 __email__ = 'bibi21000@gmail.com'
 
 
-def get_parser_import(description='Description'):
-    """Text import parser
-    """
-    parser = get_parser(description=description)
-    parser.add_argument('--delete_cache', help="Delete file in text_cache", action='store_true')
-    parser.add_argument('textfile', nargs=1, help="The file to import in text store")
-    return parser
-
-def main_import():
-    parser = get_parser_import()
-    args = parser.parse_args()
-    sourcedir, builddir = parser_makefile(args.docdir)
+@cli.command()
+@click.option('--delete/--no-delete', default=True, help="Delete file in text_cache")
+@click.argument('textfile', default=None)
+@click.pass_obj
+def store(common, delete, textfile):
+    """Import text in store"""
+    sourcedir, builddir = parser_makefile(common.docdir)
     with docutils_namespace():
         app = Sphinx(
             srcdir=sourcedir,
@@ -46,7 +42,7 @@ def main_import():
         print('Plugin text is not enabled')
         sys.exit(1)
 
-    with open(args.textfile[0], 'r') as f:
+    with open(textfile, 'r') as f:
         text = f.read()
 
     result = {
@@ -70,13 +66,13 @@ def main_import():
       "tags": None,
     }
 
-    Text.update(app, result, args.textfile[0])
+    Text.update(app, result, textfile)
 
-    storef = os.path.join(sourcedir, app.config.osint_text_store, os.path.splitext(os.path.basename(args.textfile[0]))[0] + '.json')
+    storef = os.path.join(sourcedir, app.config.osint_text_store, os.path.splitext(os.path.basename(textfile))[0] + '.json')
     with open(storef, 'w') as f:
         f.write(json.dumps(result, indent=2))
 
-    if args.delete_cache is True:
-        cachef = os.path.join(sourcedir, app.config.osint_text_cache, os.path.splitext(os.path.basename(args.textfile[0]))[0] + '.json')
+    if delete is True:
+        cachef = os.path.join(sourcedir, app.config.osint_text_cache, os.path.splitext(os.path.basename(textfile))[0] + '.json')
         if os.path.isfile(cachef):
             os.remove(cachef)